refactor: update layout component test and structure
