--- src/app/models/order-status.model.ts
@@ -1,4 +1,0 @@
-export interface OrderStatus {
-  id: number;
-  name: string;
-}
