--- src/app/models/trip-status.model.ts
@@ -1,4 +1,0 @@
-export interface TripStatus {
-  id: number;
-  name: string; // Ex: "Inițiat", "În curs", "Finalizat", "Anulat"
-}
