--- src/app/models/vehicle-type.model.ts
@@ -1,4 +1,0 @@
-export interface VehicleType {
-  id: number;
-  name: string; // Ex: 'Tractor', 'Trailer'
-}
