+++ src/app/orders/components/order-details/order-details.component.spec.ts
@@ -1,8 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { OrderDetailsComponent } from './order-details.component';
-import { OrderService } from '../../../services/order.service';
+import { OrderService } from '../../services/order.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { of } from 'rxjs';
 
 describe('OrderDetailsComponent', () => {
   let component: OrderDetailsComponent;
@@ -12,9 +11,18 @@ describe('OrderDetailsComponent', () => {
     getOrderById: (id: number) => ({
       id,
       number: 'EMAG1',
-      company: { id: 1, name: 'EMAG SRL', code: 'EMAG', contactPerson: { id: 1, fullName: 'Popescu Ion' } },
+      company: {
+        id: 1,
+        name: 'EMAG SRL',
+        code: 'EMAG',
+        contactPerson: { id: 1, fullName: 'Popescu Ion' }
+      },
       deliveryPerson: { id: 2, fullName: 'Ionescu Maria' },
-      status: { id: 1, name: 'Inițiat' },
+      status: {
+        id: 1,
+        dictionaryId: 1,
+        name: 'Inițiat'
+      },
       createdDate: new Date(),
       trip: undefined
     }),
