+++ src/app/orders/services/order.service.spec.ts
@@ -1,6 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { OrderService } from './order.service';
 import { Order } from '../models/order.model';
+import { DictionaryItem } from '../../shared/models/dictionary-item.model';
 
 describe('OrderService', () => {
   let service: OrderService;
@@ -42,6 +43,12 @@ describe('OrderService', () => {
   it('should not update a non-existent order', () => {
     const initialLength = service.getOrders().length;
 
+    const fakeStatus: DictionaryItem = {
+      id: 1,
+      name: 'Inițiat',
+      dictionaryId: 1
+    };
+
     const fakeOrder: Order = {
       id: 999,
       number: 'FAKE999',
@@ -53,12 +60,21 @@ describe('OrderService', () => {
         contactPerson: { id: 99, fullName: 'Anonim' }
       },
       deliveryPerson: { id: 98, fullName: 'Nimeni' },
-      status: { id: 1, name: 'Inițiat' },
+      status: fakeStatus,
       trip: undefined
     };
 
     service.updateOrder(fakeOrder);
     expect(service.getOrders().length).toBe(initialLength);
     expect(service.getOrderById(999)).toBeUndefined();
+  });
+
+  it('should include an order in a mock trip', () => {
+    const order = service.getOrderById(1)!;
+    service.includeInMockTrip(order);
+
+    const updatedOrder = service.getOrderById(1);
+    expect(updatedOrder?.trip).toBeDefined();
+    expect(updatedOrder?.status.name).toBe('Inclusă în cursă');
   });
 });
