+++ src/app/shared/models/dictionary-item.model.ts
@@ -1,0 +1,5 @@
+export interface DictionaryItem {
+  id: number;
+  dictionaryId: number;
+  name: string; // ex: 'Inițiat', 'Finalizat', 'Tractor', 'Trailer'
+}
