+++ src/app/shared/models/dictionary.model.ts
@@ -1,0 +1,4 @@
+export interface Dictionary {
+  id: number;
+  name: string; // ex: 'OrderStatus', 'TripStatus', 'VehicleType'
+}
